requirements.txt
requests==2.31.0
flatlib==0.3.2
pyswisseph==2.10.3.2
google-generativeai==0.3.2
